requirements.txt
PyYAML==6.0.2
pandas==2.3.1
beautifulsoup4==4.13.4
aiohttp==3.12.14
PyMuPDF==1.26.3
requests==2.32.4
transformers==4.54.0
vertexai==1.71.1
dotenv==0.9.9
llama-cpp-python==0.3.14
openai==1.98.0
evaluate==0.4.5
rouge-score==0.1.2
nltk==3.9.1
fastapi==0.116.1
uvicorn==0.35.0


